# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Continua API
  version: 0.1.0
  description: AI Agent Observability Platform API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
    description: Relative path (works with any host)
security:
  - auth0Bearer: []
  - apiKey: []
paths:
...
  /api/traces/{id}:
    get:
      operationId: getTrace
      summary: Get a trace by ID
      tags: [Traces]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SelectedProjectId'
      responses:
        '200':
          description: Trace details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceDetail'
        '401':
          description: Unauthorized - missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    SelectedProjectId:
      name: project_id
      in: query
      description: Required when authenticating with an Auth0 bearer token. Ignored when the request is already project-scoped by API key.
      schema:
        type: string
        format: uuid
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    auth0Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0 access token for debugger operator authentication
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    EngineProjectionState:
      type: string
      enum: [up_to_date, catching_up, summary_only, journal_expired]
...
    EngineRunStatus:
      type: string
      enum: [QUEUED, RUNNING, WAITING, SUSPENDED, COMPLETED, FAILED, CANCELLED, TERMINATED, CONTINUED_AS_NEW]
    EngineTraceInfo:
      type: object
      required: [run_id, definition_name, definition_version, projection_state]
      properties:
        run_id:
          type: string
          format: uuid
        definition_name:
          type: string
        definition_version:
          type: string
        projection_state:
          $ref: '#/components/schemas/EngineProjectionState'
        parent_run_id:
          type: string
          format: uuid
        root_run_id:
          type: string
          format: uuid
        child_key:
          type: string
        child_depth:
          type: integer
    EngineWaitState:
      type: object
      additionalProperties: true
      properties:
        kind:
          type: string
        activity_key:
          type: string
        activity_type:
          type: string
        timer_key:
          type: string
        due_at:
          type: string
          format: date-time
        signal_name:
          type: string
        child_key:
          type: string
    EnginePendingWork:
      type: object
      required: [pending_activity_tasks, pending_inbox_items]
      properties:
        pending_activity_tasks:
          type: integer
          format: int64
        pending_inbox_items:
          type: integer
          format: int64
...
    EngineFailureSummary:
      type: object
      required: [error_code, error_message, status]
      properties:
        error_code:
          type: string
          description: |
            Stable reserved value: `definition_version_mismatch`. Clients may
            exact-match this value while still accepting arbitrary strings.
        error_message:
          type: string
        status:
          type: string
    EngineRunSummary:
      allOf:
        - $ref: '#/components/schemas/EngineTraceInfo'
        - type: object
          required: [status, instance_key, pending_work, created_at, updated_at]
          properties:
            status:
              $ref: '#/components/schemas/EngineRunStatus'
            instance_key:
              type: string
            continued_from_run_id:
              type: string
              format: uuid
            continued_to_run_id:
              type: string
              format: uuid
            continued_from_trace_id:
              type: string
            continued_to_trace_id:
              type: string
            custom_status:
              type: object
            wait_state:
              $ref: '#/components/schemas/EngineWaitState'
            pending_work:
              $ref: '#/components/schemas/EnginePendingWork'
            result:
              description: Terminal workflow result payload when available.
            failure:
              $ref: '#/components/schemas/EngineFailureSummary'
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            completed_at:
              type: string
              format: date-time
    Trace:
      type: object
      required: [id, name, status, started_at]
      properties:
        id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        session_external_id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [RUNNING, COMPLETED, FAILED]
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        total_tokens_in:
          type: integer
        total_tokens_out:
          type: integer
        total_cost_usd:
          type: number
        error_count:
          type: integer
          description: Count of failed spans in this trace
        metadata:
          type: object
        engine:
          $ref: '#/components/schemas/EngineTraceInfo'
    TraceDetail:
      allOf:
        - $ref: '#/components/schemas/Trace'
        - type: object
          properties:
            trace_id:
              type: string
            user_id:
              type: string
            tags:
              type: array
              items:
                type: string
            environment:
              type: string
            release:
              type: string
            input:
              description: Trace input payload (any valid JSON)
            output:
              description: Trace output payload (any valid JSON)
            engine:
              $ref: '#/components/schemas/EngineRunSummary'
...